requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
pytz==2025.2
sqlparse==0.5.3
